Always reset the cursor to the beginning
The default computation mistakes a cursor at the end of the buffer
as a selection, so any update that adds more text will select text
from the old end of the buffer to the new end of the buffer.
Here, we simply set the range to the beginning to avoid this undesirable effect.