Fix imports; disable non-standard tsconfig settings
Before this commit, TypeScript users of react-monaco-editor
would need to have allowSyntheticDefaultImports to be on
in their tsconfig, or there would be TS errors due to the way
React was being imported. As far as I can tell, simply moving
those lines to `import * as React` was all that's needed to
make it all work without these settings. This should make adoption
easier for TS users.